Modify structure of files
