+++ src/paradigma/pipelines/heart_rate_pipeline.py
@@ -1,22 +1,29 @@
-from typing import Union
-from pathlib import Path
-import pandas as pd
-import os
-import numpy as np
 import json
+import numpy as np
+import os
+import pandas as pd
+from pathlib import Path
+from scipy.signal import welch
+from scipy.signal.windows import hamming, hann
+import tsdf
 from typing import List
 
-import tsdf
-
 from paradigma.constants import DataColumns
-from paradigma.config import SignalQualityFeatureExtractionConfig, SignalQualityFeatureExtractionAccConfig, SignalQualityClassificationConfig, \
-    HeartRateExtractionConfig
-from paradigma.heart_rate.feature_extraction import extract_temporal_domain_features, extract_spectral_domain_features, extract_accelerometer_feature
-from paradigma.heart_rate.heart_rate_estimation import assign_sqa_label, extract_hr_segments, extract_hr_from_segment
+from paradigma.config import SignalQualityFeatureExtractionConfig, SignalQualityFeatureExtractionAccConfig, \
+    SignalQualityClassificationConfig, HeartRateExtractionConfig
+from paradigma.feature_extraction import compute_auto_correlation, compute_dominant_frequency, compute_relative_power, \
+    compute_spectral_entropy, compute_statistics, compute_signal_to_noise_ratio
+from paradigma.pipelines.heart_rate_utils import assign_sqa_label, extract_hr_segments, extract_hr_from_segment
 from paradigma.segmenting import tabulate_windows, WindowedDataExtractor
 from paradigma.util import read_metadata, aggregate_parameter
 
-def extract_signal_quality_features(config_ppg: SignalQualityFeatureExtractionConfig, df_ppg: pd.DataFrame, config_acc: SignalQualityFeatureExtractionAccConfig, df_acc: pd.DataFrame) -> pd.DataFrame:
+
+def extract_signal_quality_features(
+        config_ppg: SignalQualityFeatureExtractionConfig, 
+        df_ppg: pd.DataFrame, 
+        config_acc: SignalQualityFeatureExtractionAccConfig, 
+        df_acc: pd.DataFrame
+    ) -> pd.DataFrame:
     """	
     Extract signal quality features from the PPG signal.
     The features are extracted from the temporal and spectral domain of the PPG signal.
@@ -68,19 +75,19 @@ def extract_signal_quality_features(config_ppg: SignalQualityFeatureExtractionCo
 
     df_features = pd.DataFrame(start_time, columns=[DataColumns.TIME])
     # Compute features of the temporal domain of the PPG signal
-    df_temporal_features = extract_temporal_domain_features(config_ppg, ppg_windowed, quality_stats=['var', 'mean', 'median', 'kurtosis', 'skewness'])
+    df_temporal_features = extract_temporal_domain_features(ppg_windowed, config_ppg, quality_stats=['var', 'mean', 'median', 'kurtosis', 'skewness'])
     
     # Combine temporal features with the start time
     df_features = pd.concat([df_features, df_temporal_features], axis=1)
 
     # Compute features of the spectral domain of the PPG signal
-    df_spectral_features = extract_spectral_domain_features(config_ppg, ppg_windowed)
+    df_spectral_features = extract_spectral_domain_features(ppg_windowed, config_ppg)
 
     # Combine the spectral features with the previously computed temporal features
     df_features = pd.concat([df_features, df_spectral_features], axis=1)
     
     # Compute periodicity feature of the accelerometer signal
-    df_accelerometer_feature = extract_accelerometer_feature(config_acc, acc_values_windowed, ppg_windowed)
+    df_accelerometer_feature = extract_accelerometer_feature(acc_values_windowed, ppg_windowed, config_acc)
 
     # Combine the accelerometer feature with the previously computed features
     df_features = pd.concat([df_features, df_accelerometer_feature], axis=1)
@@ -88,15 +95,15 @@ def extract_signal_quality_features(config_ppg: SignalQualityFeatureExtractionCo
     return df_features
 
 
-def extract_signal_quality_features_io(input_path: Union[str, Path], output_path: Union[str, Path], config_ppg: SignalQualityFeatureExtractionConfig, config_acc: SignalQualityFeatureExtractionAccConfig) -> pd.DataFrame:
+def extract_signal_quality_features_io(input_path: str | Path, output_path: str | Path, config_ppg: SignalQualityFeatureExtractionConfig, config_acc: SignalQualityFeatureExtractionAccConfig) -> pd.DataFrame:
     """
     Extract signal quality features from the PPG signal and save them to a file.
 
     Parameters
     ----------
-    input_path : Union[str, Path]
+    input_path : str | Path
         The path to the directory containing the preprocessed PPG and accelerometer data.
-    output_path : Union[str, Path]
+    output_path : str | Path
         The path to the directory where the extracted features will be saved.
     config_ppg: SignalQualityFeatureExtractionConfig
         The configuration for the signal quality feature extraction of the ppg signal.
@@ -125,7 +132,11 @@ def extract_signal_quality_features_io(input_path: Union[str, Path], output_path
     return df_windowed
 
 
-def signal_quality_classification(df: pd.DataFrame, config: SignalQualityClassificationConfig, path_to_classifier_input: Union[str, Path]) -> pd.DataFrame:
+def signal_quality_classification(
+        df: pd.DataFrame, 
+        config: SignalQualityClassificationConfig,
+        path_to_classifier_input: str | Path
+    ) -> pd.DataFrame:
     """
     Classify the signal quality of the PPG signal using a logistic regression classifier. A probability close to 1 indicates a high-quality signal, while a probability close to 0 indicates a low-quality signal.
     The classifier is trained on features extracted from the PPG signal. The features are extracted using the extract_signal_quality_features function.
@@ -138,7 +149,7 @@ def signal_quality_classification(df: pd.DataFrame, config: SignalQualityClassif
         The DataFrame containing the PPG features and the accelerometer feature for signal quality classification.
     config : SignalQualityClassificationConfig
         The configuration for the signal quality classification.
-    path_to_classifier_input : Union[str, Path]
+    path_to_classifier_input : str | Path
         The path to the directory containing the classifier.
 
     Returns
@@ -168,7 +179,12 @@ def signal_quality_classification(df: pd.DataFrame, config: SignalQualityClassif
     return df[[DataColumns.PRED_SQA_PROBA, DataColumns.PRED_SQA_ACC_LABEL]]  # Return only the relevant columns, namely the predicted probabilities for the PPG signal quality and the accelerometer label
 
 
-def signal_quality_classification_io(input_path: Union[str, Path], output_path: Union[str, Path], path_to_classifier_input: Union[str, Path], config: SignalQualityClassificationConfig) -> None:
+def signal_quality_classification_io(
+        input_path: str | Path, 
+        output_path: str | Path, 
+        path_to_classifier_input: str | Path, 
+        config: SignalQualityClassificationConfig
+    ) -> None:
     
     # Load the data
     metadata_time, metadata_values = read_metadata(input_path, config.meta_filename, config.time_filename, config.values_filename)
@@ -177,7 +193,11 @@ def signal_quality_classification_io(input_path: Union[str, Path], output_path: 
     df_sqa = signal_quality_classification(df_windowed, config, path_to_classifier_input)
 
 
-def estimate_heart_rate(df_sqa: pd.DataFrame, df_ppg_preprocessed: pd.DataFrame, config: HeartRateExtractionConfig) -> pd.DataFrame:  
+def estimate_heart_rate(
+        df_sqa: pd.DataFrame, 
+        df_ppg_preprocessed: pd.DataFrame, 
+        config: HeartRateExtractionConfig
+    ) -> pd.DataFrame:  
     """
     Estimate the heart rate from the PPG signal using the time-frequency domain method.
 
@@ -286,19 +306,22 @@ def aggregate_heart_rate(hr_values: np.ndarray, aggregates: List[str] = ['mode',
     return aggregated_results
 
 
-def aggregate_heart_rate_io(full_path_to_input: Union[str, Path], full_path_to_output: Union[str, Path], aggregates: List[str] = ['mode', '99p']) -> None:
+def aggregate_heart_rate_io(
+        full_path_to_input: str | Path, 
+        full_path_to_output: str | Path, 
+        aggregates: List[str] = ['mode', '99p']
+    ) -> None:
     """
     Extract heart rate from the PPG signal and save the aggregated heart rate estimates to a file.
 
     Parameters
     ----------
-    input_path : Union[str, Path]
+    input_path : str | Path
         The path to the directory containing the heart rate estimates.
-    output_path : Union[str, Path]
+    output_path : str | Path
         The path to the directory where the aggregated heart rate estimates will be saved.
     aggregates : List[str]
         The list of aggregation methods to be used for the heart rate estimates. The default is ['mode', '99p'].
-
     """
 
     # Load the heart rate estimates
@@ -312,3 +335,193 @@ def aggregate_heart_rate_io(full_path_to_input: Union[str, Path], full_path_to_o
     # Save the aggregated heart rate estimates
     with open(full_path_to_output, 'w') as json_file:
         json.dump(df_hr_aggregates, json_file, indent=4)
+
+
+def extract_temporal_domain_features(
+        ppg_windowed: np.ndarray, 
+        config: SignalQualityFeatureExtractionConfig, 
+        quality_stats: List[str] = ['mean', 'std']
+    ) -> pd.DataFrame:
+    """
+    Compute temporal domain features for the ppg signal. The features are added to the dataframe. Therefore the original dataframe is modified, and the modified dataframe is returned.
+
+    Parameters
+    ----------
+    ppg_windowed: np.ndarray
+        The dataframe containing the windowed accelerometer signal
+
+    config: SignalQualityFeatureExtractionConfig
+        The configuration object containing the parameters for the feature extraction
+
+    quality_stats: list, optional
+        The statistics to be computed for the gravity component of the accelerometer signal (default: ['mean', 'std'])
+    
+    Returns
+    -------
+    pd.DataFrame
+        The dataframe with the added temporal domain features.
+    """
+    
+    feature_dict = {}
+    for stat in quality_stats:
+        feature_dict[stat] = compute_statistics(ppg_windowed, stat)
+    
+    feature_dict['signal_to_noise'] = compute_signal_to_noise_ratio(ppg_windowed)  
+    feature_dict['auto_corr'] = compute_auto_correlation(ppg_windowed, config.sampling_frequency)
+    return pd.DataFrame(feature_dict)
+
+
+def extract_spectral_domain_features(
+        ppg_windowed: np.ndarray,
+        config: SignalQualityFeatureExtractionConfig, 
+    ) -> pd.DataFrame:
+    """
+    Calculate the spectral features (dominant frequency, relative power, and spectral entropy)
+    for each segment of a PPG signal using a single Welch's method computation. The features are added to the dataframe. 
+    Therefore the original dataframe is modified, and the modified dataframe is returned.
+
+    Parameters
+    ----------
+    ppg_windowed: np.ndarray
+        The dataframe containing the windowed ppg signal
+
+    config: SignalQualityFeatureExtractionConfig
+        The configuration object containing the parameters for the feature extraction
+
+    Returns
+    -------
+    pd.DataFrame
+        The dataframe with the added spectral domain features.
+    """
+    d_features = {}
+
+    window = hamming(config.window_length_welch, sym = True)
+
+    freqs, psd = welch(
+        ppg_windowed,
+        fs=config.sampling_frequency,
+        window=window,
+        noverlap=config.overlap_welch_window,
+        nfft=max(256, 2 ** int(np.log2(ppg_windowed.shape[1]))),
+        detrend=False,
+        axis=1
+    )
+
+    n_samples_window = ppg_windowed.shape[1]
+
+    # Calculate each feature using the computed PSD and frequency array
+    d_features['f_dom'] = compute_dominant_frequency(freqs, psd)
+    d_features['rel_power'] = compute_relative_power(freqs, psd, config)
+    d_features['spectral_entropy'] = compute_spectral_entropy(psd, n_samples_window) 
+
+    return pd.DataFrame(d_features)
+
+
+def extract_acc_power_feature(
+        f1: np.ndarray, 
+        PSD_acc: np.ndarray, 
+        f2: np.ndarray, 
+        PSD_ppg: np.ndarray
+    ) -> np.ndarray:
+    """
+    Extract the accelerometer power feature in the PPG frequency range.
+
+    Parameters
+    ----------
+    f1: np.ndarray
+        The frequency bins of the accelerometer signal.
+    PSD_acc: np.ndarray
+        The power spectral density of the accelerometer signal.
+    f2: np.ndarray
+        The frequency bins of the PPG signal.
+    PSD_ppg: np.ndarray
+        The power spectral density of the PPG signal.
+
+    Returns
+    -------
+    np.ndarray
+        The accelerometer power feature in the PPG frequency range
+    """
+    
+    # Find the index of the maximum PSD value in the PPG signal
+    max_PPG_psd_idx = np.argmax(PSD_ppg, axis=1)
+    max_PPG_freq_psd = f2[max_PPG_psd_idx]
+    
+    # Find the neighboring indices of the maximum PSD value in the PPG signal
+    df_idx = np.column_stack((max_PPG_psd_idx - 1, max_PPG_psd_idx, max_PPG_psd_idx + 1))    
+    
+    # Find the index of the closest frequency in the accelerometer signal to the first harmonic of the PPG frequency
+    corr_acc_psd_fh_idx = np.argmin(np.abs(f1[:, None] - max_PPG_freq_psd*2), axis=0)
+    fh_idx = np.column_stack((corr_acc_psd_fh_idx - 1, corr_acc_psd_fh_idx, corr_acc_psd_fh_idx + 1))   
+    
+    # Compute the power in the ranges corresponding to the PPG frequency
+    acc_power_PPG_range = (
+        np.trapz(PSD_acc[np.arange(PSD_acc.shape[0])[:, None], df_idx], f1[df_idx], axis=1) +
+        np.trapz(PSD_acc[np.arange(PSD_acc.shape[0])[:, None], fh_idx], f1[fh_idx], axis=1)
+    )
+
+    # Compute the total power across the entire frequency range
+    acc_power_total = np.trapz(PSD_acc, f1)
+    
+    # Compute the power ratio of the accelerometer signal in the PPG frequency range
+    acc_power_ratio = acc_power_PPG_range / acc_power_total
+    
+    return acc_power_ratio
+
+def extract_accelerometer_feature(
+        acc_windowed: np.ndarray, 
+        ppg_windowed: np.ndarray,
+        config: SignalQualityFeatureExtractionAccConfig
+    ) -> pd.DataFrame:
+    """
+    Extract accelerometer features from the accelerometer signal in the PPG frequency range.
+    
+    Parameters
+    ----------    
+    acc_windowed: np.ndarray
+        The dataframe containing the windowed accelerometer signal
+
+    ppg_windowed: np.ndarray
+        The dataframe containing the corresponding windowed ppg signal
+    
+    config: SignalQualityFeatureExtractionAccConfig
+        The configuration object containing the parameters for the feature extraction
+
+    Returns
+    -------
+    pd.DataFrame
+        The dataframe with the relative power accelerometer feature.
+    """
+    
+    d_acc_feature = {}
+
+    window_acc = hann(config.window_length_welch_acc, sym = True)
+    window_ppg = hann(config.window_length_welch_ppg, sym = True)
+
+    freqs_acc, psd_acc = welch(
+        acc_windowed,
+        fs=config.sampling_frequency,
+        window=window_acc,
+        noverlap=config.overlap_welch_window_acc,
+        nfft=config.nfft_acc,
+        detrend=False,
+        axis=1
+    )
+
+    psd_acc = np.sum(psd_acc, axis=2)  # Sum the PSDs of the three axes
+
+    freqs_ppg, psd_ppg = welch(
+        ppg_windowed,
+        fs=config.sampling_frequency_ppg,
+        window=window_ppg,
+        noverlap=config.overlap_welch_window_ppg,
+        nfft=config.nfft_ppg,
+        detrend=False,
+        axis=1
+    )
+
+    d_acc_feature['acc_power_ratio'] = extract_acc_power_feature(freqs_acc, psd_acc, freqs_ppg, psd_ppg)
+
+    return pd.DataFrame(d_acc_feature)
+
+
