+++ src/paradigma/pipelines/tremor_pipeline.py
@@ -3,13 +3,14 @@ import json
 import pandas as pd
 import numpy as np
 from pathlib import Path
-from typing import Union
+from scipy import signal
 from scipy.stats import gaussian_kde
 
 from paradigma.classification import ClassifierPackage
 from paradigma.constants import DataColumns
 from paradigma.config import TremorFeatureExtractionConfig, TremorDetectionConfig, TremorAggregationConfig
-from paradigma.tremor.feature_extraction import extract_spectral_domain_features
+from paradigma.feature_extraction import compute_power_in_bandwidth, compute_mfccs, compute_total_power, \
+    extract_frequency_peak, extract_tremor_power
 from paradigma.segmenting import tabulate_windows, WindowedDataExtractor
 from paradigma.util import get_end_iso8601, write_df_data, read_metadata, aggregate_parameter
 
@@ -61,14 +62,15 @@ def extract_tremor_features(df: pd.DataFrame, config: TremorFeatureExtractionCon
     df_features = pd.DataFrame(start_time, columns=[DataColumns.TIME])
     
     # transform the signals from the temporal domain to the spectral domain and extract tremor features
-    df_spectral_features = extract_spectral_domain_features(config, windowed_gyro)
+    df_spectral_features = extract_spectral_domain_features(windowed_gyro, config)
 
     # Combine spectral features with the start time
     df_features= pd.concat([df_features, df_spectral_features], axis=1)
 
     return df_features
 
-def extract_tremor_features_io(input_path: Union[str, Path], output_path: Union[str, Path], config: TremorFeatureExtractionConfig) -> None:
+
+def extract_tremor_features_io(input_path: str | Path, output_path: str | Path, config: TremorFeatureExtractionConfig) -> None:
     # Load data
     metadata_time, metadata_values = read_metadata(input_path, config.meta_filename, config.time_filename, config.values_filename)
     df = tsdf.load_dataframe_from_binaries([metadata_time, metadata_values], tsdf.constants.ConcatenationType.columns)
@@ -94,7 +96,7 @@ def extract_tremor_features_io(input_path: Union[str, Path], output_path: Union[
     write_df_data(metadata_time, metadata_values, output_path, 'tremor_meta.json', df_windowed)
 
 
-def detect_tremor(df: pd.DataFrame, config: TremorDetectionConfig, full_path_to_classifier_package: Union[str, Path]) -> pd.DataFrame:
+def detect_tremor(df: pd.DataFrame, config: TremorDetectionConfig, full_path_to_classifier_package: str | Path) -> pd.DataFrame:
     """
     Detects tremor in the input DataFrame using a pre-trained classifier and applies a threshold to the predicted probabilities.
 
@@ -115,7 +117,7 @@ def detect_tremor(df: pd.DataFrame, config: TremorDetectionConfig, full_path_to_
     config : TremorDetectionConfig
         Configuration object containing settings for tremor detection, including the frequency range for rest tremor.
 
-    full_path_to_classifier_package : Union[str, Path]
+    full_path_to_classifier_package : str | Path
         The path to the directory containing the classifier file, threshold value, scaler parameters, and other necessary input
         files for tremor detection.
 
@@ -170,7 +172,7 @@ def detect_tremor(df: pd.DataFrame, config: TremorDetectionConfig, full_path_to_
     return df
 
 
-def detect_tremor_io(input_path: Union[str, Path], output_path: Union[str, Path], path_to_classifier_input: Union[str, Path], config: TremorDetectionConfig) -> None:
+def detect_tremor_io(input_path: str | Path, output_path: str | Path, path_to_classifier_input: str | Path, config: TremorDetectionConfig) -> None:
     
     # Load the data
     metadata_time, metadata_values = read_metadata(input_path, config.meta_filename, config.time_filename, config.values_filename)
@@ -191,7 +193,7 @@ def detect_tremor_io(input_path: Union[str, Path], output_path: Union[str, Path]
     write_df_data(metadata_time, metadata_values, output_path, 'tremor_meta.json', df)
 
 
-def aggregate_tremor(df: pd.DataFrame, config: TremorAggregationConfig):
+def aggregate_tremor(df: pd.DataFrame, config: TremorAggregationConfig) -> dict:
     """
     Quantifies the amount of tremor time and tremor power, aggregated over all windows in the input dataframe.
     Tremor time is calculated as the number of the detected tremor windows, as percentage of the number of windows 
@@ -209,8 +211,9 @@ def aggregate_tremor(df: pd.DataFrame, config: TremorAggregationConfig):
 
     Returns
     -------
-    A json file with the aggregated tremor time and tremor power measures, as well as the total number of windows
-    available in the input dataframe, and the number of windows at rest.
+    dict
+        A dictionary with the aggregated tremor time and tremor power measures, as well as the total number of windows
+        available in the input dataframe, and the number of windows at rest.
 
     Notes
     -----
@@ -263,7 +266,7 @@ def aggregate_tremor(df: pd.DataFrame, config: TremorAggregationConfig):
     return d_aggregates
 
 
-def aggregate_tremor_io(path_to_feature_input: Union[str, Path], path_to_prediction_input: Union[str, Path], output_path: Union[str, Path], config: TremorAggregationConfig) -> None:
+def aggregate_tremor_io(path_to_feature_input: str | Path, path_to_prediction_input: str | Path, output_path: str | Path, config: TremorAggregationConfig) -> None:
     
     # Load the features & predictions
     metadata_time, metadata_values = read_metadata(path_to_feature_input, config.meta_filename, config.time_filename, config.values_filename)
@@ -286,3 +289,83 @@ def aggregate_tremor_io(path_to_feature_input: Union[str, Path], path_to_predict
     # Save output
     with open(output_path / "tremor_aggregates.json", 'w') as json_file:
         json.dump(d_aggregates, json_file, indent=4)
+
+
+def extract_spectral_domain_features(data: np.ndarray, config) -> pd.DataFrame:
+    """
+    Compute spectral domain features from the gyroscope data.
+
+    This function computes Mel-frequency cepstral coefficients (MFCCs), the frequency of the peak, 
+    the tremor power, and the non-tremor power based on the total power spectral density of the windowed gyroscope data.
+
+    Parameters
+    ----------
+    data : numpy.ndarray
+        A 2D numpy array where each row corresponds to a window of gyroscope data.
+    config : object
+        Configuration object containing settings such as sampling frequency, window type, 
+        and MFCC parameters.
+    
+    Returns
+    -------
+    pd.DataFrame
+        The feature dataframe containing the extracted spectral features, including 
+        MFCCs, the frequency of the peak, the tremor power and non-tremor power for each window.
+    """
+
+    # Initialize a dictionary to hold the results
+    feature_dict = {}
+
+    # Initialize parameters
+    sampling_frequency = config.sampling_frequency
+    segment_length_s = config.segment_length_s
+    overlap_fraction = config.overlap_fraction
+    spectral_resolution = config.spectral_resolution
+    window_type = 'hann'
+
+    # Compute the power spectral density
+    segment_length_n = sampling_frequency * segment_length_s
+    overlap_n = segment_length_n * overlap_fraction
+    window = signal.get_window(window_type, segment_length_n, fftbins=False)
+    nfft = sampling_frequency / spectral_resolution
+
+    freqs, psd = signal.welch(
+        x=data, 
+        fs=sampling_frequency, 
+        window=window, 
+        nperseg=segment_length_n,
+        noverlap=overlap_n, 
+        nfft=nfft, 
+        detrend=False, 
+        scaling='density',
+        axis=1
+    )
+
+    # Compute total power in the PSD
+    total_psd = compute_total_power(psd)
+
+    # Compute the MFCC's
+    config.mfcc_low_frequency = config.fmin_mfcc
+    config.mfcc_high_frequency = config.fmax_mfcc
+    config.mfcc_n_dct_filters = config.n_dct_filters_mfcc
+    config.mfcc_n_coefficients = config.n_coefficients_mfcc
+
+    mfccs = compute_mfccs(
+        total_power_array=total_psd,
+        config=config,
+        multiplication_factor=1
+    )
+
+    # Combine the MFCCs into the features DataFrame
+    mfcc_colnames = [f'mfcc_{x}' for x in range(1, config.mfcc_n_coefficients + 1)]
+    for i, colname in enumerate(mfcc_colnames):
+        feature_dict[colname] = mfccs[:, i]
+
+    # Compute the frequency of the peak, non-tremor power and tremor power
+    feature_dict['freq_peak'] = extract_frequency_peak(freqs, total_psd, config.fmin_peak, config.fmax_peak)
+    feature_dict['low_freq_power'] = compute_power_in_bandwidth(freqs, total_psd, config.fmin_low_power, config.fmax_low_power, 
+                                                                include_max=True, spectral_resolution=config.spectral_resolution, 
+                                                                cumulative_sum_method='sum')
+    feature_dict['tremor_power'] = extract_tremor_power(freqs, total_psd, config.fmin_tremor_power, config.fmax_tremor_power)
+
+    return pd.DataFrame(feature_dict)
